+++ src/ezdxf/entities/underlay.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2022 Manfred Moitzi
+# Copyright (c) 2019-2023 Manfred Moitzi
 # License: MIT License
 from __future__ import annotations
 from typing import TYPE_CHECKING, Union, Iterable, Optional
@@ -14,11 +14,15 @@ from ezdxf.lldxf.attributes import (
 from ezdxf.lldxf.const import SUBCLASS_MARKER, DXF2000, DXFTypeError
 from ezdxf.lldxf import const
 from ezdxf.lldxf.tags import Tags
-from ezdxf.math import NULLVEC, Z_AXIS, UVec
-from .dxfentity import base_class, SubclassProcessor
+from ezdxf.math import NULLVEC, Z_AXIS, UVec, Matrix44, Vec3
+from .dxfentity import base_class, SubclassProcessor, DXFEntity
 from .dxfgfx import DXFGraphic, acdb_entity
 from .dxfobj import DXFObject
 from .factory import register_entity
+from ezdxf.math.transformtools import (
+    InsertTransformationError,
+    InsertCoordinateSystem,
+)
 
 if TYPE_CHECKING:
     from ezdxf.document import Drawing
@@ -111,8 +115,10 @@ class Underlay(DXFGraphic):
         self._boundary_path: list[UVec] = []
         self._underlay_def: Optional[UnderlayDefinition] = None
 
-    def raw_copy(self):
-        raise DXFTypeError("Copying of underlay not supported.")
+    def copy_data(self, entity: DXFEntity) -> None:
+        assert isinstance(entity, Underlay)
+        entity._boundary_path = list(self._boundary_path)
+        entity._underlay_def = self._underlay_def
 
     def load_dxf_attribs(
         self, processor: Optional[SubclassProcessor] = None
@@ -149,6 +155,11 @@ class Underlay(DXFGraphic):
         super().post_load_hook(doc)
         db = doc.entitydb
         self._underlay_def = db.get(self.dxf.get("underlay_def_handle", None))  # type: ignore
+
+    def post_bind_hook(self):
+        assert isinstance(self.dxf.handle, str)
+        if isinstance(self._underlay_def, UnderlayDefinition):
+            self._underlay_def.append_reactor_handle(self.dxf.handle)
 
     def export_entity(self, tagwriter: AbstractTagWriter) -> None:
         """Export entity specific data as DXF tags."""
@@ -260,10 +271,40 @@ class Underlay(DXFGraphic):
         del self._boundary_path
         super().destroy()
 
+    def transform(self, m: Matrix44) -> Underlay:
+        """Transform UNDERLAY entity by transformation matrix `m` inplace.
+
+        Unlike the transformation matrix `m`, the UNDERLAY entity can not
+        represent a non-orthogonal target coordinate system and an
+        :class:`InsertTransformationError` will be raised in that case.
+
+        """
+        dxf = self.dxf
+        source_system = InsertCoordinateSystem(
+            insert=Vec3(dxf.insert),
+            scale=(dxf.scale_x, dxf.scale_y, dxf.scale_z),
+            rotation=dxf.rotation,
+            extrusion=dxf.extrusion,
+        )
+        try:
+            target_system = source_system.transform(m)
+        except InsertTransformationError:
+            raise InsertTransformationError(
+                "UNDERLAY entity can not represent a non-orthogonal target coordinate system."
+            )
+        dxf.insert = target_system.insert
+        dxf.rotation = target_system.rotation
+        dxf.extrusion = target_system.extrusion
+        dxf.scale_x = target_system.scale_factor_x
+        dxf.scale_y = target_system.scale_factor_y
+        dxf.scale_z = target_system.scale_factor_z
+        self.post_transform(m)
+        return self
+
 
 @register_entity
 class PdfUnderlay(Underlay):
-    """DXF PDFUNDERLAY entity - BricsCAD export PDFREFERENCE"""
+    """DXF PDFUNDERLAY entity - BricsCAD exports PDFREFERENCE"""
 
     DXFTYPE = "PDFUNDERLAY"