+++ NEWS.md
@@ -7,6 +7,7 @@ Version 1.0.2b0 - dev
 
 - Release notes: https://ezdxf.mozman.at/release-v1-0.html
 - NEW: `Drawing.validate()` also prints report of resolved issues
+- NEW: copy and transform support for `PDFUNDERLAY`, `DWFUNDERLAY` and `DGNUNDERLAY`
 - BUGFIX: fix ownership of sub-entities of `INSERT` and `POLYLINE` entities
 
 Version 1.0.1 - 2023-01-14